Revamped hashing documentation
New algorithm to keep a tight circle around the center-point, without relying on a bounding-box. We get to cut down on the number of accidental document fetches. Users can tweak the request for their use case: they can set the radius + set the precision, which controls the size of the search cells.

The function is also being updated to work with more than one geopoint in a document, and the geohash can be accessed in the map '__geohash__'.

All that said, and especially with the new circle-algo implemented in the request, the client code is even simpler and doesn't warrant a custom-operator anymore.